make profile icons loadable
